Added reviews for two other matmul submissions.
As part of the matmul assignment, we have to review submissions by two
other groups. Ideally, we're supposed to submit these reviews as pull
requests to the submitter's  repo, but there are some technical issues
with doing that. For now, I've included the files here.

The two groups we graded can be found here:

- https://github.com/yalcinozhabes/matmul-/blob/master/Group007.pdf
- https://github.com/jyk46/matmul-/blob/master/Group008.pdf